pm: move stats code into its own file
Stats are an optional feature, moving it into a separate source file
improves code readability.
